Improve EnergyGridCache performance in the 'steady state' situations
A 'steady state' I'm defining as the network either has an excess of power and
full storage or not having enough power and empty storage. And of course the
network having neither energy input or output. This is optimized for because
it's both easier to optimize and it's the usual state for networks. 'Unstable
states' where the network is filling up storage or emptying storage will
natually fill or empty storage completely and enter a 'steady state.' The only
case this isn't true is with an intermittent power source like solar, but an
intermittent power source will most likely have an external power storage that'll
prevent the network from seeing the power fluctuations.

This makes two main changes
1. The EnergyGridCache stores the current amount and max power in all the cells
as an internal variable and only writes to the energy cells when the Grid itself
is ticked (ie once a tick no matter how many times energy is input or removed
from the network)
2. The EnergyGridCache only writes the amount to the cells if it significantly
changed from the last tick

The first change isn't much but it allows the second change to work, basically
optimizing when the network is in a 'steady state' In these states the
EnergyGridCache won't write the energy amounts to the cells since next tick
it'll most likely fill up or empty again. Instead it makes every cell completely
full or empty once when it fill or empties and won't write until something
changes.

These changes aren't actually that important due to how AE2 handles
energy cells, but it's still useful to force connected energy cells to max which is
nice if you're trying to fill up some energy cells to take somewhere.