Changes UVLModelLoader to use a custom model implementation
This basically renders things with set brightness with bloom since they're
supposed to emitting light and it makes sense for them to have a bloom
effect. This is achieved by reimplementing Forge's VanillaModelWrapper (big
thanks to Forge, most things here are stolen from them) and when the model is
asked for it's quads it picks which quad set to return based off the current
BlockRenderLayer.

This implementation has the benefit of natively working with other mods using
AE2's UVLMap system, but comes with the downside that blocks need to be told to
render in the Bloom layer. So this ends up breaking other mod's blocks that have
light maps, but works perfectly fine for parts since those are handled by AE2
(and this commit adds bloom support to cable parts)