Seperate out Controllers into ChannelSource's to allow addons to create custom controllers
This changes the PathGridCache from hardcoding TileController and it instead
uses IChannelSources. IChannelSources validate themselves and allow addons to
arbitrarily decides what's is and isn't allowed for the controller. This even
lets multiple controllers exist on one network, if a custom implementation allowed it, but no
testing on that functionality has been done.